plus_gopay_links/requirements.txt
requests
curl-cffi
PySocks
grpcio
grpcio-tools
